debug 확인

+++ spring12rest/src/main/java/com/kh/spring12/dto/CountryDto.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class CountryDto {
-	private int countryNo;
+	private long countryNo;
 	private String countryName;
 	private String countryCapital;
 	private long countryPopulation;
